Rename package di -> dagger
