+++ presentation/src/main/java/by/grodno/vasili/presentation/dagger/ViewModelKey.java
@@ -1,4 +1,4 @@
-package by.grodno.vasili.presentation.di;
+package by.grodno.vasili.presentation.dagger;
 
 import android.arch.lifecycle.ViewModel;
